+++ linode/resource_linode_instance_test.go
@@ -781,7 +781,7 @@ func TestAccLinodeInstance_tag(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckLinodeInstanceExists(resName, &instance),
 					resource.TestCheckResourceAttr(resName, "tags.#", "1"),
-					resource.TestCheckResourceAttr(resName, "tags.0", "tf_test"),
+					resource.TestCheckResourceAttr(resName, "tags.4106436895", "tf_test"),
 				),
 			},
 			// Apply updated tags
@@ -790,8 +790,8 @@ func TestAccLinodeInstance_tag(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckLinodeInstanceExists(resName, &instance),
 					resource.TestCheckResourceAttr(resName, "tags.#", "2"),
-					resource.TestCheckResourceAttr(resName, "tags.0", "tf_test"),
-					resource.TestCheckResourceAttr(resName, "tags.1", "tf_test_2"),
+					resource.TestCheckResourceAttr(resName, "tags.4106436895", "tf_test"),
+					resource.TestCheckResourceAttr(resName, "tags.2667398925", "tf_test_2"),
 				),
 			},
 		},