+++ linode/resource_linode_nodebalancer.go
@@ -88,6 +88,12 @@ func resourceLinodeNodeBalancer() *schema.Resource {
 					},
 				},
 			},
+			"tags": &schema.Schema{
+				Type:        schema.TypeSet,
+				Elem:        &schema.Schema{Type: schema.TypeString},
+				Optional:    true,
+				Description: "An array of tags applied to this object. Tags are for organizational purposes only.",
+			},
 		},
 	}
 }
@@ -143,6 +149,7 @@ func resourceLinodeNodeBalancerRead(d *schema.ResourceData, meta interface{}) er
 	d.Set("region", nodebalancer.Region)
 	d.Set("ipv4", nodebalancer.IPv4)
 	d.Set("ipv6", nodebalancer.IPv6)
+	d.Set("tags", nodebalancer.Tags)
 	d.Set("client_conn_throttle", nodebalancer.ClientConnThrottle)
 	d.Set("created", nodebalancer.Created.Format(time.RFC3339))
 	d.Set("updated", nodebalancer.Updated.Format(time.RFC3339))
@@ -172,6 +179,13 @@ func resourceLinodeNodeBalancerCreate(d *schema.ResourceData, meta interface{}) 
 		Label:              &label,
 		ClientConnThrottle: &clientConnThrottle,
 	}
+
+	if tagsRaw, tagsOk := d.GetOk("tags"); tagsOk {
+		for _, tag := range tagsRaw.(*schema.Set).List() {
+			createOpts.Tags = append(createOpts.Tags, tag.(string))
+		}
+	}
+
 	nodebalancer, err := client.CreateNodeBalancer(context.Background(), createOpts)
 	if err != nil {
 		return fmt.Errorf("Error creating a Linode NodeBalancer: %s", err)
@@ -194,14 +208,23 @@ func resourceLinodeNodeBalancerUpdate(d *schema.ResourceData, meta interface{}) 
 		return fmt.Errorf("Error fetching data about the current NodeBalancer: %s", err)
 	}
 
-	if d.HasChange("label") || d.HasChange("client_conn_throttle") {
+	if d.HasChange("label") || d.HasChange("client_conn_throttle") || d.HasChange("tags") {
 		label := d.Get("label").(string)
 		clientConnThrottle := d.Get("client_conn_throttle").(int)
+
 		// @TODO nodebalancer.GetUpdateOptions, avoid clobbering client_conn_throttle
 		updateOpts := linodego.NodeBalancerUpdateOptions{
 			Label:              &label,
 			ClientConnThrottle: &clientConnThrottle,
 		}
+
+		tags := []string{}
+		for _, tag := range d.Get("tags").(*schema.Set).List() {
+			tags = append(tags, tag.(string))
+		}
+
+		updateOpts.Tags = &tags
+
 		if nodebalancer, err = client.UpdateNodeBalancer(context.Background(), nodebalancer.ID, updateOpts); err != nil {
 			return err
 		}